Change TankLimits to be unsigned integers

+++ heating_service/HeatingService.Contracts/HeatPump/TankLimitsResponse.cs
@@ -1,7 +1,7 @@
 ﻿namespace HeatingService.Contracts.HeatPump;
 
 public record TankLimitsResponse(
-  float LowerTankMinimum,
-  float LowerTankMaximum,
-  float UpperTankMinimum,
-  float UpperTankMaximum);
+  UInt32 LowerTankMinimum,
+  UInt32 LowerTankMaximum,
+  UInt32 UpperTankMinimum,
+  UInt32 UpperTankMaximum);
